core/requirements.txt
# Django Core
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Environment Variables
python-decouple==3.8
django-environ==0.11.2

# Multi-tenancy
django-tenants==3.6.1

# Filtering, Search & Pagination
django-filter==24.1

# Validation
drf-spectacular==0.27.1

# Task Queue (optional - for async jobs)
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0

# Utilities
Pillow==10.2.0
python-slugify==8.0.4

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==24.2.0
isort==5.13.2
